pause: add ``sv_allow_team_change_during_pause``
Not being able to change teams while the game is paused
is in teeworlds since forever.
There has to be a reason for it so I added a config to be able to revert
it to the old behavior.

But by default players can now switch teams during a paused game.

This allows timeouted players to reclaim their spot without having
to start the game again.
But it also allows trolls to occupy that spot.
So for serious games with moderation such as tournaments
where admins can use the ``set_team`` command it will
be useful to disallow joining the game during pause.

Closed #244