pause: do not spam new ddnet clients in chat (x players not ready)
The chat message was only a hack because ddnet clients used to
force open the scoreboard which hides the broadcast.

Most players use recent ddnet clients. So the hack is now
deprecated.

Old clients still receive the backwards compatible chat message
so they know why the game is paused. But they also get informed
that this message can be hidden with a client update.

New clients do no longer receive any chat message. Even if they
manually open the scoreboard and do not see the broadcast.
Players should realize quickly that they do have to stop holding
the tab key to see the broadcast. And even if not they will
be able to see the checkmarks.

Not showing it in chat anymore was an intentional decision.
It is common practice to chat and stalk scoreboard stats during
the pause. So we do not want to pollute the chat in that case.

Closed #251