refactor: favor ddnet's ``IsTeamPlay()`` if possible
In teeworlds it is called ``IsTeamplay()``
In ddnet it is called ``IsTeamPlay()``
In ddnet-insta there are both but only ``IsTeamplay()`` is const.

After https://github.com/ddnet/ddnet/pull/9656 both will be const
then we can remove ``IsTeamplay()`` which creates an annoying
diff to teeworlds but at least we match ddnet.