rounds: do not finish warmup when all players are ready
I assume this is some sort of feature that was introduced in 0.7
that there is no need to play warmup when all players pressed the ready
bind.

But then the name is not fitting anymore. Warmup should be for
warming up and not for waiting for players.
Having a timer that suddenly stops is just weird.

csgo/cs2 solved this nicely by saying it is waiting for players with a
timer. And then reducing the timer as soon as everyone is ready and
then being officially in warmup state.

I do not think there is a need for that in teeworlds.
Either way the current implementation was just weird.
So this commit now removes all early endings for warmups.
Users always have to wait for the entire timer to go to zero.
If it is too slow they can always send a follow up vote with
a shorter timer.

So this removes a "feature" that exists in vanilla teeworlds.
But I would rather call it a bug https://github.com/teeworlds/teeworlds/issues/3271
or unfinished/unclear feature at best

Closed #136