rainbow: send 0.7 rainbow for topscorer and scoreboard
Do not distance network clips rainbow if the receiving player
has the scoreboard open.
Or if the rainbowed player is a top scorer in a non team mode.
Which places him at the bottom right into the hud.

Now the rainbow clipping that is done for performance in 0.7
should be invisible to the user unless the entire server has
score 0. In that case the "top scorer" in the bottom right
will not update.