stats!: start using win_points and support scorelimit in zCatch
See #276, #191 and #274

Reaching the scorelimit in zCatch now ends the round.
But nobody gains a win in the database as it does not fullfill
the requirements of only killing the entire server.

Maybe in the future these wins will also be tracked.

BREAKING CHANGE:

Also zCatch now uses the "win_points" instead of the "points"
column for points that are given to the winner.

To migrate your old database manual action is required.
You can move the old points values to the win_points column.
And then you could also update the old points column but the value is
not really clear. It could be either reset to 0 or set to the amount
of kills. Or could also be kept as the old value.

To migrate old points to the new column you can use the following
queries depending on wether you used grenade, laser or both:

```SQL
UPDATE zcatch_grenade SET win_points = points;
UPDATE zcatch_laser SET win_points = points;
```