docs: recommend to disable ddrace timeout protection
In pvp modes the long ddrace timeout protection just
hides the fact that players are no longer actively playing
and enables farming of afk kills.

Also restoring a old tee just to get the round score is not
as desirable as restoring progress in a ddrace map.
Especially if this restoring only works for 0.6 and for 0.7
players it just causes duplicated tees on rejoin.