weapons: fix accidental shot fired during manual respawn
**The problem:**

There are situations where players can respawn faster if they
fire during the death screen. For example when the death was
caused by the world such as a spike. Even if the fire key is
held during the death screen until the tee is spawned there
is no actual shot fired in vanilla gametypes because players
spawn with the gun which is a non automatic weapon.
Meaning holding down the fire key does not keep shooting.
Only fresh clicks cause shots to be fired.

In ddnet-insta there are modes where players can spawn with
automatic weapons such as grenade and laser. Affected modes
include for example fng, gdm, gctf, ictf and zcatch and more.

So here pressing the fire key for a bit too long during death
screen causes one shot to be fired directly on spawn.
This shot can be seen as undesired for two reasons:
- If it hits the kill might be unfair for the killed tee because it was
  basically a random shot appearing out of nowhere.
  And the kill could also be unwanted by the killer because the real
  intention was respawning not shooting. This can cause for example
  unwanted team or self damage.
- If the shot does not hit the spawning player is punished with worse
  hit accuracy stats. So if a player aims for nice all time or round
  stats with as many shots as possible being enemy hits the respawning
  becomes an annoying dance.

**The fix:**

The fix turned out to be a bit more complicated because I wanted to make
sure it covers all cases as good as possible.
To fix it I made all weapons behave like non automatic weapons for half
a second after spawning.
This way it feels really close to vanilla because holding down the fire
button does not shoot even if it stays held until the tee is fully
spawned. But if the player releases the fire button and manually
reclicks a shot can be fired instantly on spawn. I wanted to make sure
that there is no delay here if someone really tries to shoot on spawn.
Otherwise players can be spawn trapped badly if the map allows it.

But if a player fully holds down fire forever. It will also activate the
automatic mode after half a second.

Sometimes players stuck their input to hold fire to farm kills while
being afk. This is still possible. But if they spawn next to a active
human the active human being able to reclick will have the upper hand
here. Which I think is really smooth.

Closed #289