config: add ``sv_teamdamage``
It does not include the modern teeworlds fix that tracks teams
in projectiles. See upstream commit here:
https://github.com/teeworlds/teeworlds/commit/a6852413bd9b1dfd7698a65e18c02745f3cb82f6

Which means that you can if team damage is off fire a shot as team blue.
Switch team to red before the shot lands. And then it can damage a blue
player even if the shot was technically launched from a blue tee.

This can for sure be considerd wrong behavior because you can launch
shots by spawning in the enemy base and you can not be damaged by those
players while you launch the shot.
But in reality team switching is a messy process and slow. So there is
no real competitive advantage of abusing this "bug".

So since this edge case does not cause actual harm. And the code
required to fix it causes additional complexity. I will intentionally
not fix this.

Closed #303