ddrteams: limit /team cmd to spectators only
This should fix all shenanigans that would give a tactical
advantage in game by joining and leaving a team.

Requiring players to first join spectators before joining
a ddrace team makes sure they can not use ddrace teams
to unfreeze in fng. And also ensures they can not hide
during a round even while unfrozen. For example as flagger
in CTF or in a last man standing mode such as zCatch.

It is now the opposite of ddnet which forces you to be in
game to join a team.
Which unlocks another feature: being able to join teams
while joining the game is not possible.

That is nice for players that want to wonder in the map
while the in game slot limit is full. Or while a zCatch
round is running.

Closed #335