clang: remove nullptr check to please clang
Seems like checking m_pController for NULL
tripped clang! Like I made it aware that in other
cases it is non NULL checked.

I assume it is safe to not check it.
As clang pointed out if someone manages to execute
a chat command during reload or server launch somehow
while the controller is not there it would crash in a
bunch of other places too anyways.