fng: fix spikes not counting without hooking the frozen tee
If you just freeze an enemy with your gun (was only broken in laser fng)
and that enemy directly falls into a spike without you hooking it.
It wont be counted as your kill.

This commit fixes that. By setting the freezer also as the last toucher.
The last toucher is the one who gets the kill on spike.

While the battlefly and block modes were added the laser weapon
was excluded from setting the last toucher in b62960e6fe2f2cab1631f485778fe51891ed52dd
because the laser does technically not move a tee at all so in battlefly
and block which also uses touchers you do not get a kill if you laser
someone.

I also thought about falling back to original freezer id when the
last toucher is not there. But that seems flawed because now the fall
back depends on the last toucher disconnecting. Also it probably
breaks that team mates can rescue their peers by resetting the last
toucher with a hook.

The same should be done for any freezing weapon I guess.
But ddnet removed the freeze hammer so we do not have another freezing
weapon right now.

Closed #354