fix: dont allow slash cmd votes when bang commands are off
cmds such as /swap, /shuffle and /swap_random
are basically just aliases for the bang commands
which are chat commands to call votes

The config ``sv_bang_commands`` now technically has the wrong name
because it applies to chat command based votes with a bang and with a
slash. But I don't think adding a new config for chat cmd votes is any
better. Or even needed. They are probably always turned on together with
bang commands anyways.