stats: add /round_top and ``sv_print_round_stats``
Show best players in a few categories.
This is the first draft. The message should still be polished.
On round end the scoreboard will be open so the slignment looks
horrible with the line wraps.
Also there are so many lines and dumping that all into chat is
a big mess. Could use the motd instead because it at least
does not get scrolled away. But the motd can not be shown on round
end because the scoreboard would block it.

See #376